feat: Add include_contents option to LlmAgentConfig to control inclusion of previous event contents in LLM requests
This CL introduces a new configuration option include_contents to the LlmAgentConfig. This allows users to specify whether the contents of previous events should be included in requests sent to the underlying LLM. The option can be set to DEFAULT (include contents) or NONE (do not include contents) in the agent configuration YAML file. The LlmAgent and LlmAgentConfig classes have been updated to handle this new field, including case-insensitive parsing of the enum values from the configuration.

PiperOrigin-RevId: 802678374