fix: Return Completable from `saveArtifact` in `CallbackContext`
The return type of the artifactService is Single, meaning it doesn't get scheduled to execute until it is subscribed. Today it's not subscribed. This is a bug leads to saveArtifact never executes.

Changing the return type to Completable allows clients subscribe to it and fits naturally with the async pattern of Orcas APIs.

PiperOrigin-RevId: 820279963